Updated Claus' Extension Mechanisms and Coverted to adoc
I added three placeholder images and an adoc file "extension_mechanisms.adoc" with Claus' doc with some edits. I left questions and notes all as NOTEs and gave a few suggestions on questions. It's ready for Chuck to review and merge with overall structure. I kept it in a separate document for clarity.